+++ python/ray/rllib/bc/experience_dataset.py
@@ -10,6 +10,16 @@ import numpy as np
 
 class ExperienceDataset(object):
     def __init__(self, dataset_path):
+        """Create dataset of experience to imitate.
+
+        Parameters
+        ----------
+        dataset_path:
+          Path of file containing the database as pickled list of trajectories,
+          each trajectory being a list of steps,
+          each step containing the observation and action as its first two elements.
+          The file must be available on each machine used by a BCEvaluator.
+        """
         self._dataset = list(itertools.chain.from_iterable(pickle.load(open(dataset_path, "rb"))))
 
     def sample(self, batch_size):
